break another circdep
